enrichment-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
requests==2.31.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
httpx==0.25.2
fake-useragent==1.4.0
beautifulsoup4==4.12.2
aiohttp==3.9.1
asyncio==3.4.3
nest-asyncio==1.5.8